Adjusted a large number of files to work with VS2008 compiler

+++ src/libs/opmapcontrol/src/mapwidget/gpsitem.cpp
@@ -210,7 +210,7 @@ namespace mapcontrol
     double GPSItem::Distance3D(const internals::PointLatLng &coord, const int &altitude)
     {
        return sqrt(pow(internals::PureProjection::DistanceBetweenLatLng(this->coord,coord)*1000,2)+
-       pow(this->altitude-altitude,2));
+       pow(static_cast<float>(this->altitude-altitude),2));
 
     }
     void GPSItem::SetUavPic(QString UAVPic)
